+++ src/background/installer.ts
@@ -17,7 +17,7 @@
 
 import { locator as serviceLocator } from "@/background/locator";
 import { type Runtime } from "webextension-polyfill";
-import { Events, reportEvent } from "@/telemetry/events";
+import { reportEvent } from "@/telemetry/reportEvent";
 import { initTelemetry } from "@/background/telemetry";
 import { getUID } from "@/background/messenger/api";
 import { allowsTrack, DNT_STORAGE_KEY } from "@/telemetry/dnt";
@@ -29,6 +29,7 @@ import { isEmpty } from "lodash";
 import { expectContext } from "@/utils/expectContext";
 import { AUTOMATION_ANYWHERE_SERVICE_ID } from "@/contrib/automationanywhere/contract";
 import { readManagedStorage } from "@/store/enterprise/managedStorage";
+import { Events } from "@/telemetry/events";
 
 const UNINSTALL_URL = "https://www.pixiebrix.com/uninstall/";
