+++ src/contentScript/sidebarController.tsx
@@ -16,7 +16,8 @@
  */
 
 import reportError from "@/telemetry/reportError";
-import { Events, reportEvent } from "@/telemetry/events";
+import { reportEvent } from "@/telemetry/reportEvent";
+import { Events } from "@/telemetry/events";
 import { expectContext } from "@/utils/expectContext";
 import sidebarInThisTab from "@/sidebar/messenger/api";
 import { isEmpty } from "lodash";
