+++ src/contrib/google/sheets/RequireGoogleApi.tsx
@@ -24,7 +24,8 @@ import initGoogle, {
 } from "@/contrib/google/initGoogle";
 import AsyncButton from "@/components/AsyncButton";
 import useUserAction from "@/hooks/useUserAction";
-import { Events, reportEvent } from "@/telemetry/events";
+import { reportEvent } from "@/telemetry/reportEvent";
+import { Events } from "@/telemetry/events";
 import { detectBrowser } from "@/vendors/mixpanel";
 
 /**
