+++ src/contrib/google/sheets/useGoogleSpreadsheetPicker.ts
@@ -20,7 +20,8 @@ import { isAuthRejectedError } from "@/contrib/google/auth";
 import { isNullOrBlank } from "@/utils";
 import { type Data, type Doc } from "@/contrib/google/sheets/types";
 import pDefer from "p-defer";
-import { Events, reportEvent } from "@/telemetry/events";
+import { reportEvent } from "@/telemetry/reportEvent";
+import { Events } from "@/telemetry/events";
 import useUserAction from "@/hooks/useUserAction";
 import { CancelError } from "@/errors/businessErrors";
 import { ensureSheetsReady } from "@/contrib/google/sheets/handlers";
