+++ src/extensionConsole/pages/deployments/DeploymentModal.tsx
@@ -22,7 +22,8 @@ import { useDispatch, useSelector } from "react-redux";
 import settingsSlice from "@/store/settingsSlice";
 import notify from "@/utils/notify";
 import { useUpdateAvailable } from "@/extensionConsole/pages/UpdateBanner";
-import { Events, reportEvent } from "@/telemetry/events";
+import { reportEvent } from "@/telemetry/reportEvent";
+import { Events } from "@/telemetry/events";
 import { selectAuth } from "@/auth/authSelectors";
 import { noop } from "lodash";
 import { selectUpdatePromptState } from "@/store/settingsSelectors";
