+++ src/extensionConsole/pages/mods/hooks/useReactivateAction.ts
@@ -19,7 +19,8 @@ import { type ModViewItem } from "@/types/modTypes";
 import { useDispatch } from "react-redux";
 import useFlags from "@/hooks/useFlags";
 import { isModComponentFromRecipe, isModDefinition } from "@/utils/modUtils";
-import { Events, reportEvent } from "@/telemetry/events";
+import { reportEvent } from "@/telemetry/reportEvent";
+import { Events } from "@/telemetry/events";
 import { push } from "connected-react-router";
 import notify from "@/utils/notify";
