+++ src/extensionConsole/pages/settings/FactoryResetSettings.tsx
@@ -23,7 +23,8 @@ import { clearPackages } from "@/baseRegistry";
 import { clearLogs, reactivateEveryTab } from "@/background/messenger/api";
 import { sessionChangesActions } from "@/store/sessionChanges/sessionChangesSlice";
 import AsyncButton from "@/components/AsyncButton";
-import { Events, reportEvent } from "@/telemetry/events";
+import { reportEvent } from "@/telemetry/reportEvent";
+import { Events } from "@/telemetry/events";
 import { useModals } from "@/components/ConfirmationModal";
 import { type Permissions } from "webextension-polyfill";
 import { selectAdditionalPermissionsSync } from "webext-additional-permissions";
