+++ src/pageEditor/EditorContent.tsx
@@ -18,7 +18,8 @@
 import React, { useEffect } from "react";
 import { useSelector } from "react-redux";
 import { selectSessionId } from "@/pageEditor/slices/sessionSelectors";
-import { Events, reportEvent } from "@/telemetry/events";
+import { reportEvent } from "@/telemetry/reportEvent";
+import { Events } from "@/telemetry/events";
 import { useGetMarketplaceListingsQuery } from "@/services/api";
 import PermissionsPane from "@/pageEditor/panes/PermissionsPane";
 import BetaPane from "@/pageEditor/panes/BetaPane";
