+++ src/pageEditor/hooks/useResetExtension.ts
@@ -24,7 +24,8 @@ import { extensionToFormState } from "@/pageEditor/starterBricks/adapter";
 import reportError from "@/telemetry/reportError";
 import { initRecipeOptionsIfNeeded } from "@/pageEditor/starterBricks/base";
 import { selectSessionId } from "@/pageEditor/slices/sessionSelectors";
-import { Events, reportEvent } from "@/telemetry/events";
+import { reportEvent } from "@/telemetry/reportEvent";
+import { Events } from "@/telemetry/events";
 import { type UUID } from "@/types/stringTypes";
 import { useAllRecipes } from "@/recipes/recipesHooks";
