+++ src/pageEditor/sidebar/InstalledEntry.tsx
@@ -41,7 +41,8 @@ import {
 import { thisTab } from "@/pageEditor/utils";
 import cx from "classnames";
 import { selectSessionId } from "@/pageEditor/slices/sessionSelectors";
-import { Events, reportEvent } from "@/telemetry/events";
+import { reportEvent } from "@/telemetry/reportEvent";
+import { Events } from "@/telemetry/events";
 import {
   selectActiveElement,
   selectActiveRecipeId,
