+++ src/pageEditor/tabs/editTab/useReportTraceError.test.tsx
@@ -25,7 +25,7 @@ import sessionSlice from "@/pageEditor/slices/sessionSlice";
 import { configureStore, type Store } from "@reduxjs/toolkit";
 import { type TraceRecord } from "@/telemetry/trace";
 import { uuidv4 } from "@/types/helpers";
-import { reportEvent } from "@/telemetry/events";
+import { reportEvent } from "@/telemetry/reportEvent";
 import { renderHook } from "@testing-library/react-hooks";
 import useReportTraceError from "./useReportTraceError";
 import { Provider } from "react-redux";
@@ -36,7 +36,7 @@ import {
 } from "@/testUtils/factories/traceFactories";
 
 // Override the manual mock to support `expect` assertions
-jest.mock("@/telemetry/events", () => ({
+jest.mock("@/telemetry/reportEvent", () => ({
   reportEvent: jest.fn(),
 }));
