+++ src/sidebar/activateRecipe/ActivateRecipeInputs.tsx
@@ -38,7 +38,8 @@ import { produce } from "immer";
 import { isDatabaseField } from "@/components/fields/schemaFields/fieldTypeCheckers";
 import { isUUID } from "@/types/helpers";
 import ServicesBody from "@/extensionConsole/pages/activateRecipe/ServicesBody";
-import { Events, reportEvent } from "@/telemetry/events";
+import { reportEvent } from "@/telemetry/reportEvent";
+import { Events } from "@/telemetry/events";
 
 type ActivateRecipeInputsProps = {
   recipe: ModDefinition;
