+++ src/starterBricks/quickBarExtension.tsx
@@ -39,7 +39,8 @@ import notify, {
   DEFAULT_ACTION_RESULTS,
   showNotification,
 } from "@/utils/notify";
-import { Events, reportEvent } from "@/telemetry/events";
+import { reportEvent } from "@/telemetry/reportEvent";
+import { Events } from "@/telemetry/events";
 import { selectEventData } from "@/telemetry/deployments";
 import { selectExtensionContext } from "@/starterBricks/helpers";
 import { type BrickConfig, type BrickPipeline } from "@/bricks/types";
